:tada: Use new sidebar in prod
