+++ aksel.nav.no/website/components/layout/sidebar/Sidebar.module.css
@@ -1,8 +1,157 @@
-.menuSection {
-  padding: var(--a-spacing-2);
-  border-radius: var(--a-border-radius-medium);
+.navElement {
+  position: relative;
+  flex-shrink: 0;
+  align-self: flex-start;
+  padding-inline: var(--ax-space-8);
+  width: 16rem;
+  margin-left: var(--ax-space-16);
+  color: var(--ax-text-neutral);
 }
 
-.menuSection[data-type="darkside"] {
-  background-color: var(--ax-bg-brand-magenta-moderate);
+.rootUl {
+  display: flex;
+  flex-direction: column;
+  gap: var(--ax-space-12);
+}
+
+.divider {
+  height: 1px;
+  width: 100%;
+  border-top: 1px solid var(--ax-border-neutral-subtle);
+}
+
+.navListGroup {
+  border-radius: var(--ax-border-radius-medium);
+
+  &[data-type="darkside"] {
+    background-color: var(--ax-bg-brand-magenta-moderate);
+  }
+}
+
+.navListGroupLabel {
+  padding-block: var(--ax-space-2);
+  padding-inline-start: var(--ax-space-8);
+  color: var(--ax-text-neutral-subtle);
+  font-weight: var(--ax-font-weight-bold);
+  text-transform: capitalize;
+}
+
+.navListSubButton {
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  align-self: stretch;
+  border-radius: var(--ax-border-radius-medium);
+  padding-block: var(--ax-space-4);
+  padding-inline: var(--ax-space-8) var(--ax-space-4);
+  min-height: 2rem;
+
+  &[aria-expanded="true"] + ul {
+    margin-bottom: var(--ax-space-6);
+  }
+
+  &:hover {
+    background-color: var(--ax-bg-neutral-moderate-hoverA);
+  }
+
+  &:focus-visible {
+    outline: 3px solid var(--ax-border-focus);
+    outline-offset: 0;
+    z-index: 1;
+  }
+
+  &[aria-expanded="false"][data-state="active"] {
+    background-color: var(--ax-bg-neutral-moderateA);
+
+    &:hover {
+      background-color: var(--ax-bg-neutral-moderate-hoverA);
+    }
+  }
+
+  [data-type="darkside"] &[aria-expanded="false"][data-state="active"] {
+    background-color: var(--ax-bg-brand-magenta-moderateA);
+  }
+
+  &[data-state="active"] {
+    font-weight: var(--ax-font-weight-bold);
+  }
+
+  [data-layout="mobile"] & {
+    padding-block: var(--ax-space-12);
+  }
+}
+
+.navListNotch {
+  &::before {
+    content: "";
+    position: absolute;
+    background-color: var(--ax-bg-brand-blue-strong-pressed);
+    width: 3px;
+    height: calc(100% - 8px);
+    left: -7px;
+    top: 50%;
+    transform: translateY(-50%);
+    border-radius: var(--ax-border-radius-full);
+  }
+
+  [data-type="darkside"] &::before {
+    background-color: var(--ax-bg-brand-magenta-strong-pressed);
+  }
+}
+
+.navListItem {
+  position: relative;
+  min-height: 1.75rem;
+}
+
+.navListItemLink {
+  border-radius: var(--ax-border-radius-medium);
+  display: block;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding-block: var(--ax-space-6);
+  padding-inline: var(--ax-space-8);
+
+  [data-layout="mobile"] & {
+    padding-block: var(--ax-space-12);
+  }
+
+  [data-state="active"] & {
+    font-weight: var(--ax-font-weight-bold);
+  }
+
+  [data-type="neutral"] [data-state="active"] & {
+    background-color: var(--ax-bg-brand-blue-moderateA);
+    color: var(--ax-text-brand-blue);
+  }
+
+  [data-type="neutral"] [data-state="active"]:hover & {
+    background-color: var(--ax-bg-brand-blue-moderate-hoverA);
+  }
+
+  [data-type="darkside"] [data-state="active"] & {
+    background-color: var(--ax-bg-brand-magenta-moderateA);
+    color: var(--ax-text-brand-magenta);
+  }
+
+  [data-type="darkside"] [data-state="active"]:hover & {
+    background-color: var(--ax-bg-brand-magenta-moderate-hoverA);
+  }
+
+  [data-state="inactive"]:hover & {
+    background-color: var(--ax-bg-neutral-moderate-hoverA);
+  }
+
+  [data-nested="true"] & {
+    padding-inline-start: var(--ax-space-20);
+  }
+
+  &:focus-visible {
+    outline: 3px solid var(--ax-border-focus);
+    outline-offset: 0;
+    z-index: 1;
+  }
 }
