+++ aksel.nav.no/website/components/layout/templates/WithSidebar.tsx
@@ -1,12 +1,12 @@
 import cl from "clsx";
 import Image from "next/legacy/image";
 import NextLink from "next/link";
-import { Box, Detail, Heading, Link } from "@navikt/ds-react";
-import SidebarOld from "@/layout/sidebar/Sidebar";
+import { Box, Detail, Heading, Link, Show } from "@navikt/ds-react";
 import { urlFor } from "@/sanity/interface";
 import { SidebarT, TableOfContentsT } from "@/types";
 import { capitalize } from "@/utils";
 import { TableOfContents } from "@/web/toc/TableOfContents";
+import { Sidebar } from "../sidebar/Sidebar";
 
 export const WithSidebar = ({
   children,
@@ -39,23 +39,9 @@ export const WithSidebar = ({
       className="min-h-screen-header"
     >
       <div className="mx-auto flex w-full max-w-screen-2xl gap-6">
-        <SidebarOld kategori={pageType.type} links={sidebar} />
-        {/* New sidebar implementation demo */}
-        {/* <Show asChild above="md">
-          <Sidebar
-            sidebarData={[
-              { label: "Grunnleggende", links: sidebar },
-              {
-                label: "Komponenter",
-                links: [
-                  { slug: "test-slug", tag: "ready", heading: "Test heading" },
-                  ...sidebar,
-                  { slug: "test-slug", tag: "ready", heading: "Test heading" },
-                ],
-              },
-            ]}
-          />
-        </Show> */}
+        <Show asChild above="md">
+          <Sidebar sidebarData={[{ label: pageType.type, links: sidebar }]} />
+        </Show>
         <main
           tabIndex={-1}
           id="hovedinnhold"
